requirements.txt
Flask
Fortuna
gunicorn
requests
python-dotenv
pymongo[srv]
python-multipart
